pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-focus-analyzer"
version = "1.0.0"
description = "Analyze any PDF through a custom focus lens using LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "pydantic>=2.7.0",
    "PyMuPDF>=1.24.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "tqdm>=4.66.0",
]

[tool.setuptools.packages.find]
include = ["infra*", "pipeline*"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
